% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distfuns.r
\name{define_xs}
\alias{define_xs}
\title{Cross Section}
\usage{
define_xs(C.height, C.width, ROB.height, ROB.width, LOB.height, LOB.width)
}
\arguments{
\item{C.width}{The channel width.
ROB.height The right bank height.
ROB.width The right bank floodplain width.
LOB.height The left bank height.
LOB.width The left bank floodplain width.}
}
\value{
an object of class 'xs'.
}
\description{
Define a rectangular compound cross section.
}

