% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distfuns.r
\name{deposit_mass}
\alias{deposit_mass}
\title{Mass Deposition}
\usage{
deposit_mass(xs, dy, decay, truncate.dist)
}
\arguments{
\item{xs}{The cross section object with mass data.}

\item{dy}{Spacing between cross section nodes.}

\item{decay}{The decay coefficient.}

\item{truncate.dist}{The maximum distance from the channel to deposit
on the banks.}
}
\value{
A dataframe of node positions and mass deposited.
}
\description{
Deposit mass using a decay function.
}

