% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distfuns.r
\name{distribute_mass}
\alias{distribute_mass}
\title{Distribute Mass Over Cross Section}
\usage{
distribute_mass(mass, xs, H, b, conc.fun = NULL, ...)
}
\arguments{
\item{mass}{The total mass of sediment to be distributed over the
cross-section.}

\item{xs}{The cross-section object.}

\item{H}{The water level relative to the channel invert elevation.}

\item{b}{The active layer height.}

\item{conc.fun}{The function describing vertical concentration 
gradient. If missing, a constant gradient is assumed.}
}
\value{
The cross-section object with mass distribution.
}
\description{
Distribute sediment mass over a rectangular compound cross-section.
}

