% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concfuns.r
\name{get_cb}
\alias{get_cb}
\title{Active Layer Concentration}
\usage{
get_cb(conc.avg, fun, H, b, ...)
}
\arguments{
\item{conc.avg}{The cross-section average concentration.}

\item{fun}{The vertical gradient function to use.}

\item{H}{The maximum water depth.}

\item{b}{The active layer height.}

\item{...}{Other arguments to pass to \code{fun}.}
}
\value{
The concentration at the active layer boundary.
}
\description{
Get the active layer concentration at height \code{b} above the bed.
}

