% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distfuns.r
\name{ob_dist}
\alias{ob_dist}
\title{Overbank Deposion}
\usage{
ob_dist(ob.width, ob.nodes, decay)
}
\arguments{
\item{ob.width}{The overbank width.}

\item{ob.nodes}{}
}
\description{
Compute overbank depositon based on an exponential decay function.
}

