% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\docType{data}
\name{ui}
\alias{ui}
\title{ConcentrationGames UI}
\format{An object of class \code{shiny.tag} of length 3.}
\usage{
ui
}
\description{
Shiny user interface for CocentrationGames
}
\keyword{datasets}

