% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concfuns.r
\name{sscformula}
\alias{barennblatt}
\alias{lane}
\alias{rouse}
\alias{sscformula}
\alias{tanaka}
\title{Suspended Sediment Vertical Gradient Formulas.}
\usage{
rouse(z, H, crb, w, k, us)

lane(z, H, crb, w, k, us)

barennblatt(z, H, crb, w, k, us)

tanaka(z, H, crb, w, k, us)
}
\arguments{
\item{z}{The depth at which to compute concentration.}

\item{H}{The maximum water depth.}

\item{crb}{The ratio of water depth to active layer depth \code{b/H}.}

\item{w}{The settling velocity.}

\item{k}{The von Karmen constant.
param us The shear velocity.}
}
\description{
Formulas for specifying the vertical gradient of suspended sediment.
}
\section{Functions}{
\itemize{
\item \code{rouse}: Rouse's formula

\item \code{lane}: Lane and Kalinske's formula

\item \code{barennblatt}: Barenblatt's formula

\item \code{tanaka}: Tanaka and Sugimoto's formula
}}

